requirements.txt
beautifulsoup4==4.11.1
heroku3==5.1.4
psutil==5.9.0
Pyrogram==2.0.16
python-dotenv==0.20.0
requests==2.27.1
shazamio==0.1.1.1
youtube_search_python==1.6.5
yt_dlp==2022.4.8
tgcrypto